Add the password token model
This will create a new relation for storing a password token that can be
used to reset a users password. The idea is that an email will be sent
to the user with a link containing the token. The user will have to
click on that link within a certain timeframe and the record will need
to be active. Once the link has been clicked the record will no longer
be active, so if the user clicks again the link will be invalid.

If the user fails to click the link within the specified timeframe, the
link will no longer be valid.