Separate searching for a scene and running it
Now if you pass in a string it'll search for the scene and return the
name of the scene and it's id. If you pass it an id then it will run the
scene, if it can find it. This is to make it easier for the Jasper
plugin to ensure it's running the right scene.

Also, tried to normalise the scene search string and the scene names so
that we're more likely to hit the correct scene with our search.